requirements.txt
# Django Backend
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9

# Celery
celery==5.3.4
redis==5.0.1

# Telegram Bot
aiogram==3.4.1
aiogram-dialog==2.1.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-django==4.7.0